new languages initiation

+++ java/factorial.java
@@ -5,10 +5,10 @@ class factorial
 {
  public static void main(String args[]) throws IOException
  {
-  int n,i,fact=1;
+  long n,i,fact=1;
   System.out.println(" Enter the n value ");
   DataInputStream dis=new DataInputStream(System.in);
-  n=Integer.parseInt(dis.readLine());
+  n=Long.parseLong(dis.readLine());
   if(n==0)
   {
    System.out.println("The Factorial of "+n+" is 1");
